split cliargs for simpler definitions
