+++ packages/cli/__tests__/main.test.ts
@@ -1,13 +1,10 @@
-import main, { run } from '../src/main';
+import main, { run } from '../src/run';
 
 test('throws error for unknown args', async () => {
   expect.assertions(1);
 
   try {
-    await run({
-      // @ts-ignore
-      command: 'foo'
-    });
+    await run('foo', {});
   } catch (error) {
     expect(error).toEqual(new Error("idk what i'm doing."));
   }
@@ -17,10 +14,7 @@ test('throws exits for caught error', async () => {
   console.log = jest.fn() as any;
   process.exit = jest.fn() as any;
 
-  await main({
-    // @ts-ignore
-    command: 'foo'
-  });
+  await main('foo', {});
 
   expect(process.exit).toHaveBeenCalledWith(1);
 });
